+++ tests/output/c-backend/skip-save-expr.c
@@ -1,4 +1,4 @@
 void func(long unsigned int arg)
 {
-  ( void ) ( __builtin_constant_p ( arg ) == 0 ? -1 : 1 );
+  ( void ) ( long unsigned int ) ( __builtin_constant_p ( arg ) == 0 ? -1 : 1 );
 }
